+++ libraries/CRC/src/CRC.cpp
@@ -1,3 +1,11 @@
+//
+//    FILE: CRC.cpp
+//  AUTHOR: Rob Tillaart
+// VERSION: 1.0.1
+// PURPOSE: Arduino library for CRC8, CRC12, CRC16, CRC32, CRC64
+//     URL: https://github.com/RobTillaart/CRC
+
+
 #include "CRC.h"
 
 uint8_t calcCRC8(
@@ -107,3 +115,7 @@ uint64_t crc64(
 {
   return calcCRC64(array, length, polynome, initial, xorOut, reverseIn, reverseOut, yieldPeriod);
 }
+
+
+//  -- END OF FILE
+