+++ libraries/CRC/src/CRC12.h
@@ -1,4 +1,12 @@
 #pragma once
+//
+//    FILE: CRC12.h
+//  AUTHOR: Rob Tillaart
+// PURPOSE: Arduino class for CRC12
+//     URL: https://github.com/RobTillaart/CRC
+//          https://github.com/RobTillaart/CRC/issues/13
+
+
 #include "CrcParameters.h"
 #include "CrcDefines.h"
 
@@ -8,14 +16,14 @@ public:
   CRC12(uint16_t polynome = CRC12_POLYNOME,
         uint16_t initial  = CRC12_INITIAL,
         uint16_t xorOut   = CRC12_XOR_OUT,
-        bool reverseIn    = CRC12_REF_IN,
-        bool reverseOut   = CRC12_REF_OUT);
+        bool reverseIn    = CRC12_REV_IN,
+        bool reverseOut   = CRC12_REV_OUT);
 
   void reset(uint16_t polynome = CRC12_POLYNOME,
              uint16_t initial  = CRC12_INITIAL,
              uint16_t xorOut   = CRC12_XOR_OUT,
-             bool reverseIn    = CRC12_REF_IN,
-             bool reverseOut   = CRC12_REF_OUT);
+             bool reverseIn    = CRC12_REV_IN,
+             bool reverseOut   = CRC12_REV_OUT);
 
   void restart();
   uint16_t calc() const;
@@ -62,3 +70,7 @@ private:
   uint16_t _crc;
   crc_size_t _count;
 };
+
+
+// -- END OF FILE --
+