+++ libraries/CRC/src/CRC16.cpp
@@ -1,3 +1,10 @@
+//
+//    FILE: CRC16.cpp
+//  AUTHOR: Rob Tillaart
+// PURPOSE: Arduino class for CRC16
+//     URL: https://github.com/RobTillaart/CRC
+
+
 #include "CRC16.h"
 #include "CrcFastReverse.h"
 
@@ -100,3 +107,7 @@ uint16_t CRC16::getCRC() const
 {
   return calc();
 }
+
+
+// -- END OF FILE --
+