+++ libraries/CRC/src/CRC8.cpp
@@ -1,3 +1,10 @@
+//
+//    FILE: CRC8.cpp
+//  AUTHOR: Rob Tillaart
+// PURPOSE: Arduino class for CRC8;
+//     URL: https://github.com/RobTillaart/CRC
+
+
 #include "CRC8.h"
 #include "CrcFastReverse.h"
 
@@ -100,3 +107,7 @@ uint8_t CRC8::getCRC() const
 {
   return calc();
 }
+
+
+// -- END OF FILE --
+