+++ libraries/CRC/src/CrcFastReverse.cpp
@@ -1,3 +1,10 @@
+//
+//    FILE: CrcFastReverse.cpp
+//  AUTHOR: vovagorodok
+// PURPOSE: Arduino class for CRC
+//     URL: https://github.com/RobTillaart/CRC
+
+
 #include "CrcFastReverse.h"
 
 uint8_t reverse8bits(uint8_t in)